task-6-am-page-object-login: better but still error and no driver window
