+++ alena_automationLessons/src/main/java/com/octonion/pageObject/BasePage.java
@@ -1,0 +1,4 @@
+package com.octonion.pageObject;
+
+public class BasePage {
+}
