+++ alena_automationLessons/src/main/java/com/octonion/pageObject/LoginPage.java
@@ -5,9 +5,10 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.PageFactory;
 
-public class LoginPage {
+public class LoginPage extends BasePage{
 
     private WebDriver driver;
+//    private emailInput;
 
     @FindBy(css = "input#email")
     private WebElement emailInput;
@@ -29,4 +30,9 @@ public class LoginPage {
     public void submit() {
         submitBtn.click();
     }
+
+     public LoginPage(WebDriver driver) {
+        this.driver=driver;
+        PageFactory.initElements(driver,this);
+     }
 }
