+++ alena_automationLessons/src/main/java/com/octonion/pageObject/MenuPanel.java
@@ -3,7 +3,7 @@ package com.octonion.pageObject;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 
-public class MenuPanel {
+public class MenuPanel extends BasePage{
     @FindBy(xpath = "//div[@class='custom-icon left-sidebar__link-icon']")
     private WebElement checkBox;
